Create PYAE.txt
This game is in VERY early development. I am still working on the various primary class the levels, story, and level scaling system. The Pyromancy branch is currently the most complete with the Spellsword class having the most completed parts. Please stay upto date as this is a constantly growing project, and I am very excited to share it with you!